rename date_picker to DatePicker and move to sharedComponents
